🔧separate WeeklyDay file
